Break Tanks to Fix Tanks
This has a few flaws to be worked out, but it is on the way to being
much better than the old version. I figure since this is going on a dev
branch the flaws are ok for now.

The way the blocks check for completion on a neighbor change was an
issue, so in order to get them formed atm, they have to be right-clicked
with a debugger.