Fixed Shovel Interaction Bug
To go into further detail: Many people complain that when clicking on chests / crafting tables / doors / etc.., the swing is nullified, and sometimes other things too (like opening a GUI, the door open sound, etc.). This proposal changes flag to false if the block clicked is NOT a farmland block, preventing any other code that may result in this weirdness from occurring.