Improved Item Rendering
Items now render in 3D, just like all other Minecraft items. Held item
rendering is a bit off, but it is a small price to pay for having fully
3D item models so easily.

Also, added in Forge's update check to the mod, so that now users can
see when their mod is out-of-date.

The update notification will show up on the newly cleaned up Mod Info
page.