JSON Networking
Now plants are automagically downloaded from the server, like mutations.
This allows for the practical usage of custom crops in a multiplayer
environment. Rendering the seeds and the plants remains an issue still,
however, as that is @InfinityRaider 's thing. The next step will be to
rework core to provide more useful validation messages.