Additional Refactoring
This is a definite improvement on the code. There is still quite a bit
of work to be done, as always...

I still have no idea as to what method should be used to prevent null
method parameters... I think I might just start all methods with
Objects.requireNonNull and go from there...

Null sure is a hard thing to handle.