Attempt To Isolate API
This commit should remove the last reference from the API code to the
actual AgriCraft code. Hopefully I can figure out how to completely
separate it into its own source set in gradle, so that any
cross-references are then made impossible. Still, this is good enough
for now.

On a side note, I did add a bit more JavaDoc, which is always good!