Right clicking crops with plants uses crop sticks
Currently right clicking a crop that has a plant will consume crop sticks from your inventory. This fixes that. I also moved the setCrossCrop(true) into the conditional as it only made sense to set it to true if a crossCrop was actually placed i.e. the stack count went down.