Fix TileEntitySprinkler's range being located at wrong offset
While creating the BlockRange in the constructor, the TE's pos is still
set to the ORIGIN, and the range variable is centered on 0,1,0. That
means that the sprinkler will proceed to water the bedrock. The pos
isn't set until Chunk.java:670 during block placement. As a workaround,
the range is being created during the update call. Needs more research
to determine the proper place for initializing that.