Fix particle texture causing FileNotFoundException
The string used in the ResourceLocation for the sprinkler water particle
was for an icon instead of the texture. As a result, the client would
throw an exception and then it would instead use the "missingo" texture.

Although this does fix the black and purple texture, all particles will
still be water colored. For example, blocks break into bursts of water
droplets. A separate commit fixes that though.