Add a debug mode for testing BlockRange
In this mode, right clicking the debugger on two blocks will make the
server attempt to overwrite the cuboid volume inbetween the two corners
defined by the positions of the blocks (inclusive).

It does that by creating a new BlockRange using the two BlockPos, and
then running a for-each loop on the range, and calling setBlockState on
the positions provided by the built in iterator.

Warning #1: Creates free wool. Totally OP.
Warning #2: Blocks in the way will be deleted and lost.
Warning #3: In case of further bugs, "in the way" might mean blocks that
            are outside the cuboid specified. Backup your worlds!