Rename IAgriCrop methods with AbstractMethodErrors
For the IAgriCrop interface, this renames the getPos and getWorld
methods so as to not match the methods inherited from TileEntity.
The only uses so far were in the mutation strategies. And the newly
named methods just call the original named methods via 'this'.

This workaround is for a re-obfuscation bug in Special Source, a
program that ForgeGradle depends on.

https://github.com/MinecraftForge/ForgeGradle/issues/205
https://github.com/md-5/SpecialSource/issues/12