Fix canGrow returning true for empty crops
Since empty crops and cross crops are logically "not mature", it is
necessary to also test for a plant/seed to determine if the BlockCrop
canGrow.

This *indirectly* resolves issue #1018.

What is happening is that although BlockCrop.onBlockActivated(),
BonemealWrapper.applyFertilizer(), TileEntityCrop.acceptsFertilizer()
all work correctly, if a crop is empty then they will return false to
each other, until the false (aka "nothing happened") reached up to
PlayerInteractionManager.processRightClickBlock(). And that method then
continues trying out other possibilities, until it finally reaches
ItemDye.onItemUse(). Since this is bonemeal, it will then also call
ItemDye.applyBonemeal().

What happens next is simple: first Forge sends out a BonemealEvent,
which AgriCraft dosen't subscribe to. Then, since BlockCrop implements
IGrowable, it will test canGrow(), which sometimes returns true, and
then it'll test canUseBonemeal(), which is always false. The weirdness
is because if a block can grow, then bonemeal will always be consumed
and the particle effect will trigger, even if the block *can't* use
bonemeal specifically. The only thing that is blocked is that the crop
isn't told to grow() meanwhile.

Don't ask me why, because I can't guess how these extra complicated
conditions would ever be useful. That's just a vanilla minecraft thing.

The reason this helps with the bug is that by correcting canGrow to be
false for empty crops, ItemDye.applyBoneMeal() will not consume any of
the bonemeal.