Feature: Already analyzed seeds update the journal
When a specimen or journal is inserted into the analyzer, such that an
analyzed seed is present at the same time as a journal that lacks that
plant, then the plant will be added to the journal. This makes the
behavior more intuitive, and less roundabout. Credit for pointing out
this possibility goes to @Dilnu.