SpreadStrategy stat calculation uses all neighbors
Copies the list of neighbors before removing the immature ones, so that
the full list can be passed to calculateSpreadStats. This way the
otherCropsAffectStatsNegatively config option will have the full effect.
Immature plants are counted as invalid parents in the calculations.

This commit also adds the @Nonnull annotations to the executeStrategy
method, in order to please the IDE.