Add new DebugMode for testing IGrowable blocks
This mode makes it easier to see in game how the three methods of the
IGrowable interface function. The debugger calls all three methods
(unlike regular bonemeal) and prints out the two booleans to make
visible the slight oddities of the vanilla system and method names.

Also, being able to call grow despite the denials of the first two
methods is a good way to test how the target block responds to mods
that don't follow the conventions of vanilla.

Unlike bonemeal, no particles are generated. This is to make it easier
to determine if the target block is emitting or animating on its own.
The chat feedback will be indicative of if grow was called or not.

A quick summary of vanilla: canGrow's return value indicates if the
bonemeal should be consumed by this action. Mature wheat will always
return false and end the process, while saplings always return true.
If it continues, then the caller (ie Item.Dye) will next call
canUseBonemeal. That name is unfortunately misleading. The result
indicates if grow should be called next. Immature wheat always returns
true, trusting that well behaved bonemeal will have stopped if asked
by canGrow. Saplings will return either true or false, depending on
random chance. Having two methods like this allows vanilla plants to
prevent the consumption of bonemeal, and also have a non-100% chance
for growth, while still allowing outside methods to force the growth
reliably when needed.

UPDATE: (2017-10-18)

After cherry-picking this commit, I cleaned up the exact phrasings of
the concise notes, plus expanded the comments documenting the source
and in this message. Hopefully intentions will make sense to others.