Add more tests and feedback to onApplySeeds
When applying seeds to a crop, it now separately checks soil, light,
and the other conditions, and prints specific responses into the
player's chat buffer. This should make it easier for players to
understand what is going on, and also if something is actually broken.

Additionally, now the crop blocks the attempt only if the soil is
incorrect. If the light level is wrong, or there are additional unmet
requirements (e.g. the ore blocks for the resource crops), then it
warns the player, but allows the seed to be planted.

This should make the process more intuitive, and allow the player to
more easily experiment with getting conditions perfect for a plant.
Blocking because of wrong soil makes sense anyhow because it cannot
be easily replaced without breaking the sticks.