Add a template for new issues
It should help users organize their thoughts, and structure their
submission. It covers the basic questions, but also suggests ways to
make the report more useful. Finally, it tries to encourage both
searching for past issues, and also making the submission easy to search
for in the future.

I reviewed and copied from several different templates to start off,
then rephrased the headings and prompts. So most of the current version
is original content now.