Bump version to a21 for next development
The unofficial a20 has been out long enough that I want to avoid any
possible miscommunications about which one people are testing. There was
also an a21-unfinished made for Discord and hosted on my fork, but I'm
not worried about it, plus I am dropping the unfinished label this time.

Consequentially, I moved the new entries in the most recent changelog so
that they are under a21 instead of a20. I'm counting this is a small
lesson in avoiding directly committing to the master branch. Sorry.