Fix ItemCrop blocking right clicks
By running first before the block, the ItemCrop would prevent the block
from getting activated by returning a value of EnumActionResult.FAIL.
Since the code assumes that right clicking the top of any block is an
attempt to place a crop, this would prevent players from making a cross
crop while targetting the top face, or from opening chests from above
while holding crop sticks.

The other benefit of this change is that onItemFirstUse is depreciated,
and might get removed in the future. So the code is now more standard.