Make ItemMagnifyingGlass run server-side
The method getLightFromNeighbors does not reflect the setting of the sun
when run on the client thread. So, as a stop gap, I've switched the
execution to be server-side instead. This means switching which method
gets overridden, because onItemFirstUse does not get run on the server.
To get the old behavior back (item taking priority over block), crouch
before right clicking with the magnifying glass.

Added bonus is that onItemFirstUse was depreciated anyhow, and it is
better to avoid using it.