Make crops not drop seeds when harvested
This fixes a gameplay balance issue, because it made it trivial to
duplicate a plant with its current stats. Instead, once the player has a
plant with stats that they like, they have to wait for it to spread, or
use the Clipper item.

This reverts the change in behavior introduced during the refactor on
May 24th, 2017. My assumption is that it was an oversight made during
the consolidation of code within TileEntityCrop.

I'm not proud of the style of my fix, but the variations match up well
with the three different kinds of usages.

https://github.com/AgriCraft/AgriCraft/commit/e7c4f0e78d174638b9a33d0ab45e280ed9b738c0#diff-e92d41d0d84f12f89141e0d86116150bR410