Fix for issue #1113
The sprinkler not working is due to the above channel not refreshing connections when the sprinkler is placed on it. If a channel is placed next to the one on top of sprinkler this will cause a connections refresh and the sprinkler will start working. This change makes the sprinkler update the above channel as it is placed, making it work as intended.