Forge maven, fixed peripheral (#1407)
Updated link for forge maven.
Changed the peripheral block, it now extends the seed analyzer block.
Also changed the analyzer block class, so that it can be used in this way.

The container class and the entity class of the peripheral block are all extending the seed analyzer classes, but the block class does not. Because of this, it lacks a getDrops method, and does not drop the journal when harvested. It also lacked the hardness property, and on the client you could "destroy" it with one click, but it was not actually destroyed.